cmg7-24

+++ app/Http/Controllers/Home/FavoriteController.php
@@ -10,44 +10,63 @@ class FavoriteController extends Controller
     //收藏商品页面
     public function index()
     {
+        // 判断用户是否登录
     	if (!session('user')) {
-    		return redirect('home/login/index');
+    		return redirect('home/login/index')->with(['info' => '请登录']);
     	}
     	// 查询数据
-    	$data = \DB::table('favorite')->where('uid', session('user')->id)->first();
+    	$res = \DB::table('favorite')->where('uid', session('user')->id)->first();
 
-    	if ($data) {
+    	if ($res) {
     		// 查询数据
-    		$data = \DB::table('favorite')->where('uid', session('user')->id)->get();
+    		$data = \DB::table('favorite')->where('uid', session('user')->id)->simplePaginate(3);
+            // 查询总条数
+            $max = \DB::table('favorite')->where('uid', session('user')->id)->count();
+            // 进一取整
+            $max = ceil($max/3);
 
     		// 遍历获取到的数据
 	    	foreach ($data as $key => $value) 
 	    	{
 	    		// 去数据库查询数据
-	    		$data = \DB::table('shop')->where('id', $value->sid)->get();
+	    		$datc = \DB::table('shop')->where('id', $value->sid)->get();
 	    		// 把数据存放到空数组里
-	    		$date[] = $data;
+	    		$date[] = $datc;
 	    	}
+
+            // 把数据发送到
+            return view('home/favorite/favorite', ['date' => $date, 'data' => $data, 'max' => $max]);
+
     	}else{
-    		return view('home.favorite.favorite');
+
+            // 引入页面
+    		return view('home/favorite/favorite');
     	}
 
     	
 
-    	return view('home.favorite.favorite', ['date' => $date]);
+    	
     }
 
     public function getajax(Request $request)
     {
+        // 判断是否登录
+        if (!session('user')) 
+        {
+            return redirect()->with(['info' => '请登录']);
+        }
     	// 取出商品id
     	$sid = $request->input('sid');
+        
     	// 取出用户id
     	$data['uid'] = session('user')->id;
+
     	// 取出商品id
     	$data['sid'] = $sid;
+
     	// 去数据库查询
     	$res = \DB::table('favorite')->where('sid', $sid)->first();
-    	// dd($res);
+    	
     	// 判断是否能查到结果
     	if($res)
     	{
@@ -71,6 +90,11 @@ class FavoriteController extends Controller
     // 删除收藏商品
     public function delete($id)
     {
+        // 判断用户是否登录
+        if(!session('user')) 
+        {
+            return redirect()->with(['info' => '请登录']);
+        }
     	// 删除收藏商品
     	$res = \DB::table('favorite')->where('uid', session('user')->id)->where('sid', $id)->delete();
 
@@ -78,6 +102,7 @@ class FavoriteController extends Controller
     	if ($res) 
     	{
     		return redirect('home/favorite/index');
+
     	}else{
     		return back();
     	}
